MyService going to delete

+++ src/main/java/org/wtm/web/user/service/impl/DefaultMyPageService.java
@@ -2,7 +2,11 @@ package org.wtm.web.user.service.impl;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -30,6 +34,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class DefaultMyPageService implements MyPageService {
@@ -44,9 +49,7 @@ public class DefaultMyPageService implements MyPageService {
     private final TicketHistoryPurchaseRepository ticketHistoryPurchaseRepository;
     private final TicketHistoryUsageRepository ticketHistoryUsageRepository;
     private final ReviewScoreRepository reviewScoreRepository;
-
-    @Value("${image.upload-profile-dir}")
-    private String uploadDir;
+    private final ReviewImgRepository reviewImgRepository;
 
     private final PasswordEncoder passwordEncoder;
 
@@ -61,7 +64,6 @@ public class DefaultMyPageService implements MyPageService {
     }
 
     public UserResponseDto getMySettings(Long id) {
-
         User user = userRepository.findUserById(id);
 
         return UserResponseDto.builder()
@@ -69,6 +71,7 @@ public class DefaultMyPageService implements MyPageService {
                 .name(user.getName())
                 .address(user.getAddress())
                 .phone(user.getPhone())
+                .profilePicture(user.getProfilePicture())
                 .build();
     }
 
@@ -84,9 +87,11 @@ public class DefaultMyPageService implements MyPageService {
 
             // 프로필 사진 업로드 처리
             MultipartFile file = userUpdateDto.getProfilePicture();
-            if (file != null) {  // 파일이 null이 아니고 비어있지 않으면
-                String savedFileUrl = uploadService.uploadFile(file, uploadDir);
-                existedUser.updateProfilePicture(savedFileUrl);
+            if (file != null && !file.isEmpty()) {  // 파일이 null이 아니고 비어있지 않으면
+                String savedFileUrl = uploadService.uploadFile(file, "users");
+                if (savedFileUrl != null) {
+                    existedUser.updateProfilePicture(savedFileUrl);
+                }
             }
 
             // 비밀번호 인코딩 처리
@@ -106,6 +111,7 @@ public class DefaultMyPageService implements MyPageService {
 
         // (1) 구입 티켓 수량을 ticketId로 그룹화하여 Map에 저장
         purchasedTickets.forEach(purchaseTicket -> {
+            if (purchaseTicket.getTicket() != null && purchaseTicket.getTicket().getStore() != null) {
             Long ticketId = purchaseTicket.getTicket().getId();
             // 아래 코드는 ticketId로 된 key가 없을 시, 해당 데이터를 key로 그리고 value를 ticketSummaryDto 객체로 넣어주는 것
             TicketSummaryDto ticketSummaryDto = ticketMap.computeIfAbsent(ticketId, k ->
@@ -116,9 +122,9 @@ public class DefaultMyPageService implements MyPageService {
                             0L,
                             0L,
                             0.0
-                    ));
+                            ));
             ticketSummaryDto.addPurchasedTickets(purchaseTicket.getAmount());
-
+    }
         });
 
         // (2) 사용 티켓 수량을 ticketId로 Map에 추가
@@ -142,7 +148,15 @@ public class DefaultMyPageService implements MyPageService {
         Map<Long, Double> reviewAvgMap = storeIds.stream()
                 .collect(Collectors.toMap(
                         storeId -> storeId,
-                        reviewRepository::calculateAvgByStoreId
+                        storeId -> {
+                            Double avg = reviewRepository.calculateAvgByStoreId(storeId);
+                            if(avg != null){
+                                avg = Math.round(avg * 10) / 10.0;
+                            }else{
+                                avg = 0.0;
+                            }
+                            return avg;
+                        }
                 ));
 
         // Step 4: store 정보 가져오기 및 최종 JSON 생성
@@ -169,51 +183,15 @@ public class DefaultMyPageService implements MyPageService {
         return ticketListInfo;
     }
 
-    @Transactional
-    public boolean useMyTicket(TicketUsageDto ticketUsageDto) {
-        Integer purchasedTicketAmount =
-                ticketHistoryPurchaseRepository.getPurchaseAmountByUserId(ticketUsageDto.getUserId());
-        if(purchasedTicketAmount != null && purchasedTicketAmount > 0){
-            TicketHistoryUsage ticketHistoryUsage = TicketHistoryUsage.builder()
-                    .user(userRepository.findById(ticketUsageDto.getUserId()).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다.")))
-                    .ticket(ticketRepository.findById(ticketUsageDto.getTicketId()).orElseThrow(() -> new IllegalArgumentException("티켓 정보를 찾을 수 없습니다.")))
-                    .amount(ticketUsageDto.getAmount())
-                    .build();
-
-            // 기존 엔티티가 존재하지 않을 경우에만 저장
-            if (ticketHistoryUsage.getId() == null) {
-                ticketHistoryUsageRepository.save(ticketHistoryUsage);
-                return true;
-            }else {
-                System.out.println(ticketHistoryUsage.getId() + "는 이미 존재합니다.");
-                return false;
-            }
-        }
-        System.out.println("사용할 티켓이 존재하지 않습니다");
-        return false;
-    }
-
-    @Transactional
-    public boolean purchaseMyTicket(TicketPurchaseDto ticketPurchaseDto) {
-
-        TicketHistoryPurchase ticketHistoryPurchase = TicketHistoryPurchase.builder()
-                .user(userRepository.findById(ticketPurchaseDto.getUserId())
-                        .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다.")))
-                .ticket(ticketRepository.findById(ticketPurchaseDto.getTicketId())
-                        .orElseThrow(() -> new IllegalArgumentException("티켓을 찾을 수 없습니다.")))
-                .amount(ticketPurchaseDto.getAmount())
-                .build();
-        ticketHistoryPurchaseRepository.save(ticketHistoryPurchase);
-
-        return true;
-    }
-
-    public TicketHistoryResponseDto getMyTicketHistory(Long userId, int month, int year) {
+    public TicketHistoryResponseDto getMyTicketHistory(Long userId, int month, int year, int page, int size) {
         // 시작일과 종료일 설정
         LocalDate startDate = LocalDate.of(year, month, 1);
         LocalDate endDate = startDate.with(TemporalAdjusters.lastDayOfMonth());
         LocalDateTime startDateTime = startDate.atStartOfDay();
         LocalDateTime endDateTime = endDate.plusDays(1).atStartOfDay();
+
+        // Pageable 설정
+        Pageable pageable = PageRequest.of(page-1, size, Sort.by("regDate").ascending());
 
         // TicketHistoryUsage 데이터 가져오기
         List<TicketAllHistoryDto> usageHistory = ticketHistoryUsageRepository
@@ -226,8 +204,9 @@ public class DefaultMyPageService implements MyPageService {
                         .ticketId(usage.getTicket().getId())
                         .type("usage")
                         .regDate(usage.getRegDate())
-                        .price(usage.getAmount()*ticketRepository.findPriceById(usage.getTicket().getId()))
+                        .price(usage.getAmount() * ticketRepository.findPriceById(usage.getTicket().getId()))
                         .storeName(usage.getTicket().getStore().getName())
+                        .hasReview(usage.getReview() != null)
                         .build())
                 .collect(Collectors.toList());
 
@@ -242,39 +221,46 @@ public class DefaultMyPageService implements MyPageService {
                         .ticketId(purchase.getTicket().getId())
                         .type("purchase")
                         .regDate(purchase.getRegDate())
-                        .price(purchase.getAmount()*ticketRepository.findPriceById(purchase.getTicket().getId()))
+                        .price(purchase.getAmount() * ticketRepository.findPriceById(purchase.getTicket().getId()))
                         .storeName(purchase.getTicket().getStore().getName())
                         .build())
                 .collect(Collectors.toList());
 
-        // totalPurchasedPrice와 totalUsedPrice 계산
-        Long totalPurchasedPrice = purchaseHistory.stream()
-                .mapToLong(purchase -> purchase.getAmount() * ticketRepository.findPriceById(purchase.getTicketId()))
+        // 두 목록을 합치고 날짜로 정렬
+        List<TicketAllHistoryDto> combinedHistory = Stream.concat(purchaseHistory.stream(), usageHistory.stream())
+                .sorted(Comparator.comparing(TicketAllHistoryDto::getRegDate))
+                .collect(Collectors.toList());
+
+        // 페이지네이션 수동 적용
+        int start = Math.min((int) pageable.getOffset(), combinedHistory.size());
+        int end = Math.min((start + size), combinedHistory.size());
+        List<TicketAllHistoryDto> paginatedHistory = combinedHistory.subList(start, end);
+
+        // totalPurchasedPrice와 totalUsedPrice 계산 (페이지네이션과 무관하게 전체 데이터 기준)
+        Long totalPurchasedPrice = ticketHistoryPurchaseRepository.findByUserIdAndRegDateBetween(userId, startDateTime, endDateTime)
+                .stream()
+                .mapToLong(
+                        purchase -> purchase.getAmount() * ticketRepository.findPriceById(purchase.getTicket().getId())
+                )
                 .sum();
 
-        Long totalUsedPrice = usageHistory.stream()
-                .mapToLong(usage -> usage.getAmount() * ticketRepository.findPriceById(usage.getTicketId()))
+        Long totalUsedPrice = ticketHistoryUsageRepository.findByUserIdAndRegDateBetween(userId, startDateTime, endDateTime)
+                .stream()
+                .mapToLong(usage -> usage.getAmount() * ticketRepository.findPriceById(usage.getTicket().getId()))
                 .sum();
 
         // 전체 티켓 개수 계산 (purchase - usage)
         Long totalAmount = purchaseHistory.stream().mapToLong(TicketAllHistoryDto::getAmount).sum()
                 - usageHistory.stream().mapToLong(TicketAllHistoryDto::getAmount).sum();
 
-        // 두 목록을 합치고 날짜로 정렬
-        // 메서드화 시킨다면, combinedHistory 로 합체 및 정렬화 시키는거 따로 뺄것.
-        // 이름 달라야할 이유가 오로지 하나의 서비스여서그럼.
-        List<TicketAllHistoryDto> combinedHistoryForAll = Stream.concat(purchaseHistory.stream(), usageHistory.stream())
-                .sorted(Comparator.comparing(TicketAllHistoryDto::getRegDate))
-                .collect(Collectors.toList());
-
+        // 페이지네이션된 결과만 반환하지 않고 모든 데이터를 합친 후 반환
         return TicketHistoryResponseDto.builder()
-                .combinedHistory(combinedHistoryForAll)
+                .combinedHistory(paginatedHistory)
                 .totalPurchasedPrice(totalPurchasedPrice)
                 .totalUsedPrice(totalUsedPrice)
                 .totalAmount(totalAmount)
                 .build();
     }
-
 
     public TicketHistoryResponseDto getMyTicketHistoryByStore(Long userId, Long storeId, int month, int year) {
         // 시작일과 종료일 설정
@@ -371,14 +357,15 @@ public class DefaultMyPageService implements MyPageService {
             List<UserReviewDto> reviewsWithAverageScore = reviews.stream().map(review -> {
                 // @Query를 이용해 평균 점수를 한 번에 가져옴
                 Double averageScore = reviewScoreRepository.findAverageScoreByReviewId(review.getId());
+                String reviewImg = reviewImgRepository.findImgByReviewId(review.getId());
                 // 리뷰와 평균 점수를 합쳐서 반환
                 return UserReviewDto.builder()
                         .reviewId(review.getId())
                         .content(review.getContent())
+                        .reviewImgUrl(reviewImg)
                         .regDate(review.getRegDate())
                         .averageScore(averageScore != null ? averageScore : 0.0)
                         .storeName(review.getStore().getName()) // 필요한 Store 필드를 추가
-//                         .storeImageUrl()
                         .build();
             }).collect(Collectors.toList());
 
@@ -403,6 +390,7 @@ public class DefaultMyPageService implements MyPageService {
                                 .storeName(bookmark.getStore().getName())
                                 .storeOpenTime(bookmark.getStore().getOpenTime())
                                 .storeCloseTime(bookmark.getStore().getCloseTime())
+                                .storeImgUrl(bookmark.getStore().getImg())
                                 .ticketId(ticket != null ? ticket.getId() : null)
                                 .ticketPrice(ticket != null ? ticket.getPrice() : 0)
                                 .reviewAverage(reviewRepository.calculateAvgByStoreId(storeId)) // @Query를 이용한 평균 점수 계산
@@ -415,11 +403,14 @@ public class DefaultMyPageService implements MyPageService {
     }
 
     @Transactional
-    public boolean deleteMyReview(Long reviewId) {
-        Review findReview = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new EntityNotFoundException("Review not found with ID: " + reviewId));
-        reviewRepository.deleteById(reviewId);
-        return true;
+    public boolean deleteMyReview(Long reviewId, Long userId) {
+        Review review = reviewRepository.findByIdAndUserId(reviewId, userId)
+                .orElse(null);
+        if(review!=null){
+            reviewRepository.deleteById(reviewId);
+            return true;
+        }
+        return false;
     }
 
     @Transactional
@@ -428,18 +419,19 @@ public class DefaultMyPageService implements MyPageService {
                 .orElseThrow(() -> new IllegalArgumentException("Store not found with id: " + storeId));
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found with id: " + userId));
+        // 이미 존재하는지 확인
+        if (bookmarkRepository.existsByStoreIdAndUserId(storeId, userId)) {
+            return false;
+        }
 
-        Bookmark bookmark = bookmarkRepository.findByStoreIdAndUserId(storeId, userId)
-                .orElseGet(() -> Bookmark.builder()
+        Bookmark bookmark = Bookmark.builder()
                         .store(store)
                         .user(user)
-                        .build());
+                        .build();
 
-        if (bookmark.getId() == null) {
-            bookmarkRepository.save(bookmark);
-            return true;
-        }
-        return false;
+        bookmarkRepository.save(bookmark);
+        return true;
+
     }
 
     @Transactional
@@ -454,5 +446,25 @@ public class DefaultMyPageService implements MyPageService {
         return false;
     }
 
+    public TicketDto getMyTicketDetail(Long storeId, Long userId) {
+        Ticket ticket = ticketRepository.findOneByStoreId(storeId, Sort.by(Sort.Direction.DESC, "price"))
+                .orElseThrow(()-> new EntityNotFoundException("ticket이 없어요잉 ~"));
+        Bookmark bookmark = bookmarkRepository.findOneByStoreIdAndUserId(storeId, userId);
+
+        Long purchaseAmount = ticketHistoryPurchaseRepository.countByTicketIdAndUserId(ticket.getId(), userId);
+        Long usageAmount = ticketHistoryUsageRepository.countByTicketIdAndUserId(ticket.getId(), userId);
+
+        long totalAmount = (purchaseAmount != null ? purchaseAmount : 0L) - (usageAmount != null ? usageAmount : 0L);
+
+        boolean isBookmarked = bookmark != null;
+
+        return TicketDto.builder()
+                .ticketId(ticket.getId())
+                .storeName(ticket.getStore().getName())
+                .storeId(ticket.getStore().getId())
+                .isBookmarked(isBookmarked)
+                .ticketAmount(totalAmount)
+                .build();
+    }
 
 }
